Move Messages from MailData to ThreadViewModel
