+++ app/src/main/java/com/infomaniak/mail/ui/main/thread/ThreadFragment.kt
@@ -124,7 +124,7 @@ class ThreadFragment : Fragment() {
 
     private fun listenToMessages() {
         viewModel.messages.observeNotNull(this, ::displayMessages)
-        viewModel.loadMessages(navigationArgs.threadUid)
+        viewModel.listenToThread(navigationArgs.threadUid)
     }
 
     private fun displayMessages(messages: List<Message>) {
